Dev: change construct sequence in service
